requirements.txt
wheel==0.29.0
prettyconf==1.2.3
python-dateutil==2.6.0
arrow==0.10.0
maya==0.1.8
telepot==10.5

# Flask
Flask==0.12
MarkupSafe==1.0
Werkzeug==0.12.1
Jinja2==2.9.5
itsdangerous==0.24
Flask-Migrate==2.0.3
Flask-Script==2.0.5

# Database
Flask-SQLAlchemy==2.2
SQLAlchemy==1.1.6
psycopg2==2.7.1

# Testing
pytest==3.0.7
pytest-runfailed==0.6
factory-boy==2.8.1
Faker==0.7.10
ipdb==0.10.2
ipython==5.3.0
coverage==4.3.4
pytest-cov==2.4.0

# Lint and code style
flake8==3.3.0
flake8-blind-except==0.1.1
flake8-debugger==1.4.0
flake8-docstrings==1.0.3
flake8-isort==2.1.3
flake8-quotes==0.9.0
isort==4.2.5
pep8==1.7.0
pep8-naming==0.4.1
pylint==1.6.5
